+++ relay-dynamic-config/src/global.rs
@@ -183,6 +183,14 @@ pub struct Options {
     )]
     pub http_span_allowed_hosts: Vec<String>,
 
+    /// Disables Relay from sending replay-events to Snuba.
+    #[serde(
+        rename = "replay.relay-snuba-publishing-disabled",
+        deserialize_with = "default_on_error",
+        skip_serializing_if = "is_default"
+    )]
+    pub replay_relay_snuba_publish_disabled: bool,
+
     /// All other unknown options.
     #[serde(flatten)]
     other: HashMap<String, Value>,