+++ tests/integration/test_replay_recordings.py
@@ -1,6 +1,9 @@
 import zlib
 
 from sentry_sdk.envelope import Envelope, Item, PayloadRef
+import pytest
+
+from .test_replay_events import generate_replay_sdk_event
 
 
 def test_replay_recordings(mini_sentry, relay_chain):
@@ -30,16 +33,31 @@ def test_replay_recordings(mini_sentry, relay_chain):
     assert replay_recording.startswith(b"{}\n")  # The body is compressed
 
 
+@pytest.mark.parametrize(
+    "value,expected", [(None, False), (False, False), (True, True)]
+)
 def test_nonchunked_replay_recordings_processing(
-    mini_sentry, relay_with_processing, replay_recordings_consumer, outcomes_consumer
+    mini_sentry,
+    relay_with_processing,
+    replay_events_consumer,
+    replay_recordings_consumer,
+    outcomes_consumer,
+    value,
+    expected,
 ):
     project_id = 42
     org_id = 0
     replay_id = "515539018c9b4260a6f999572f1661ee"
-    relay = relay_with_processing()
+
+    if value is not None:
+        mini_sentry.global_config["options"][
+            "replay.relay-snuba-publishing-disabled"
+        ] = value
     mini_sentry.add_basic_project_config(
         project_id, extra={"config": {"features": ["organizations:session-replay"]}}
     )
+    relay = relay_with_processing()
+    replay_events_consumer = replay_events_consumer(timeout=10)
     replay_recordings_consumer = replay_recordings_consumer()
     outcomes_consumer = outcomes_consumer()
 
@@ -54,6 +72,8 @@ def test_nonchunked_replay_recordings_processing(
     )
     payload = recording_payload(b"[]")
     envelope.add_item(Item(payload=PayloadRef(bytes=payload), type="replay_recording"))
+    json_payload = generate_replay_sdk_event()
+    envelope.add_item(Item(payload=PayloadRef(json=json_payload), type="replay_event"))
 
     relay.send_envelope(project_id, envelope)
 
@@ -67,6 +87,14 @@ def test_nonchunked_replay_recordings_processing(
     assert replay_recording["retention_days"] == 90
     assert replay_recording["payload"] == payload
     assert replay_recording["type"] == "replay_recording_not_chunked"
+    assert replay_recording["relay_snuba_publish_disabled"] is expected
+
+    if value is True:
+        # Nothing produced.
+        with pytest.raises(AssertionError):
+            replay_events_consumer.get_replay_event()
+    else:
+        assert replay_events_consumer.get_replay_event() is not None
 
     outcomes_consumer.assert_empty()